Added our own custom "allocator" paradigm. This way our C++ objects are allocated using our C API. Now we don't have to check for null objects, double-check pointers on delete, etc...
FIXME: Is there a to make the "delete" operator aware of the size of the block it is about to free?
If so, we could use some allocation routines that perform much better than system malloc().   
(Real GLib has GSlice, but they are many others...)

P.S. - Would it be feasible to add support for allocating objects using a garbage collector algorithm present ?
(I'm not sure this would really increase performance, because at the moment each object would have to be a pinned root.)